requirements.txt
streamlit==1.28.0
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.10
python-dotenv==1.0.0